@@ -10,20 +10,22 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": {
+		"node": "./index.js",
+		"default": "./browser.js"
+	},
 	"engines": {
-		"node": ">=10"
+		"node": ">=12"
 	},
 	"scripts": {
-		"test": "xo && ava"
+		"//test": "xo && ava",
+		"test": "xo"
 	},
 	"files": [
 		"index.js",
 		"browser.js"
 	],
-	"exports": {
-		"node": "./index.js",
-		"default": "./browser.js"
-	},
 	"keywords": [
 		"color",
 		"colour",
@@ -47,12 +49,11 @@
 		"16m"
 	],
 	"dependencies": {
-		"has-flag": "^4.0.0"
+		"has-flag": "^5.0.0"
 	},
 	"devDependencies": {
-		"ava": "^2.4.0",
-		"import-fresh": "^3.2.2",
-		"xo": "^0.35.0"
-	},
-	"browser": "browser.js"
+		"ava": "^3.15.0",
+		"import-fresh": "^3.3.0",
+		"xo": "^0.38.2"
+	}
 }
